Add files via upload
                                                                               CORNERSTORE PRODUCT PERFORMANCE DASHBOARD

Source of data: sales report from Restaurant manager application.
  
The above dashboard shows the performance of 13 products across 4 different categories within a period of 3 years (2020-2022). To tell how products which is our focus of this analysis fared within this period, we decided to conduct a comparative analysis. Here we compared each product of focus with atleast one other product within its category. 
The line graph at the base of the dashboard represents the units each products sold in their various categories. 

COFFEE:
The Corner Store Special was compared with the Caramel Latte due to their similarity as "tasty coffee" drinks. Sales record as represented in the dashboard shows that the cornerstore special being in same price range with the caramel latte sold more in each of the 3 years and in total. Cornerstore special sold around 100 units more than caramel latte in 3 years even though the latte was first introduced. Only very popular coffee drinks like Americano and cappuccino sold more within this period. Cornerstore special is not featured on our new menu. 

MOCKTAIL: 
Virgin colada and Cinderella is not featured on our new menu. Similar comparison was carried out to know how these mocktails fared within this period. It was compared with two other mocktails present in the menu (banana pink and Christmas punch). Sales record as shown in the dashboard reveals that the Virginia colada and Cinderella sold 217 and 190 units respectively, while banana pink and Christmas punch recorded 85 and 72 units respectively within this period. 

COCKTAIL: 
Strawberry mojito and pinacolada was the focus here as they are not featured in the new menu too. These cocktails unit sales were compared with strawberry daiquiri and sex in the driveway. This was chosen as they were within same popularity range in our store. Pina colada recorded the lowest sales, however, the popularity of the cocktail makes it difficult to be left out on a cocktail menu. Strawberry mojito on the other hand sold the most units in its category. Selling almost a hundred units more than the daiquiri which is likewise a popular drink. 

BITES: 
The sales record and analysis makes it obvious that the Vegetable wrap "underperformed". Could it be because of the relatively few vegetarian clients we have? We decided to look into vegetable wrap because although it recorded low sales, its "fans" are disappointed when they come around and it's unavailable. Some Lebanese and Indian guests have the vegetable wrap as their favorites, and sometimes come for it alone. The chicken wrap recorded the most sales, next to the beef wraps, under this category.